pyproject.toml

[project]
name = "quick-tips-awscdk"
version = "0.1.0"
description = ""
authors = [
    {name = "Erik Lundevall Zara", email = "45368867+eriklz@users.noreply.github.com"},
]
dependencies = [
    "aws-cdk-lib>=2.83.1",
    "icecream>=2.1.3",
    "constructs>=10.2.50",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "pre-commit>=3.3.2",
    "ruff>=0.0.272",
    "pytest>=7.3.2",
]
